Manually disable hyperthreading when CpuOptions aren't supported
*.metal instance types don't support the use of CpuOptions. Since that
was previously the only way disabling hyperthreading was supported by
ParallelCluster, the only way to disable hyperthreading on these
instances was to use a pre- or post-install script to utilize the
approach described
[here](https://aws.amazon.com/blogs/compute/disabling-intel-hyper-threading-technology-on-amazon-linux/).

This change enables hyperthreading to be disabled on metal instance
types by running the script from the link above as part of the
configuration process.

It also avoids attempting to disable hyperthreading via either
CpuOptions or the script when an instance type's deafult
ThreadsPerCore according to EC2's DescribeInstanceTypes API has a
value of 1. Previously cluster creation would have failed in such a
case.

There are some known issues with this change:
* The function that guesses whether or not hyperthreading can be
disabled via CpuOptions assumes that it can for any non-metal instance
type with a default ThreadsPerCore of > 1. This is known to be false
for at least one case: cc2.8xlarge. There are probably others.
* In the event that DescribeInstanceTypes doesn't contain the
information, the function that returns an instance type's default
ThreadsPerCore assumes a value of either 1 or 2 depending on the
instance type's supported architectures. Currently this information is
only missing from the API for .metal instances.