Add CloudWatch Logs Dashboard
This new dashboard contains all the useful logs that are available with AWS ParallelCluster.

The code is flexible and makes it possible to add easily other logs in the future but also makes it possible to choose the filters we apply to the log (only see warnings for example by modifying the regex).

We only show the logs of the Head Node.

Signed-off-by: Alexandre Gobeaux <gobeaa@amazon.com>