Improve pcluster-config utility by removing unrelated sections from the output
# Removing unrelated sections from the output
The sections that are not part of the current cluster will be removed
by the output.

Added a method to retrieve the names of the section as represented
in the configuration file.
It is useful to remove the unrelated sections.
An alternative is to clear the entire config_parser but by removing the
sections one by one we're preserving the order of the existing ones.

# Skipping conversion
Right now we're avoiding to show the output to the user or write the file
if the conversion is not needed.
We're also exiting if the output file already exists.

# Tests
Added a unit test for an input file with unrelated sections.
Added assert in the tests to be sure the output contains the expected messages.
Removed expected_output files for tests not requiring the conversion

Signed-off-by: Enrico Usai <usai@amazon.com>