Add logic to cleanup pcluster_config into the convert utility
Before of this patch we were cleaning up the config_parser
only when executin the pcluster configure utility.

By adding the prepare_to_file function we're sure to have
the pcluster_config ready to be written in the file.

## Conversion logic
When a default is equal to default should not be written in the file.
It means: disable_hyperthreading = false, initial_count = min_count and min_count = 0
Instead enable_efa in the queue will be always written because the default is None

## Tests
Improved unit tests by covering all the enable_efa and dysable_hyperthreading cases.

Signed-off-by: Enrico Usai <usai@amazon.com>