Disallow using AMIs created by other versions of pcluster by AMI_name
As a result, a AMI created by other versions of pcluster cannot be used as a BASE_AMI in AMI creation or a custom_AMI in cluster creation. However, if the AMI_name does not contain pcluster version, the code in CLI package has to allow the AMI and leave the validation to CloudFormation and Cookbook stage.

Signed-off-by: Hanwen <hanwenli@amazon.com>