Unify dashboards and permit to disable dashboard
## Enabling/disabling the dashboard

Before this commit, we were creating two different dashboards, one for the EC2 metrics and one for the logs.
Both the dashboard were automatically created with the cluster.

Now, the two dashboards have been unified into a single one and the user can choose to enable (default behaviour) or disable it.

Changes:
* A new parameter `dashboard_settings` in the `cluster` section
* A new `dashboard` section
* A new `enable` boolean param in the `dashboard` section

## Updating the cluster

During a `pcluster update` the dashboard will be updated accordingly to the new settings.
It is also possible to enable/disable the dashboard during the update.

## Tests

* Added tests for the new cluster parameter `dashboard_settings` and for the new section `dashboard`.
* Adapted template rendering tests.

Signed-off-by: Alexandre Gobeaux <gobeaa@amazon.com>