Enable/disable dashboard + fix update + tests
Before this commit, the dashboard was automatically created with the cluster. Now, the user can choose to enable (default behaviour) or disable the dashboard. Updates are also possible: a user can decide that it does not want to have a dashboard anymore or to have one even though he did not have one when creating the cluster. Tests were done both for the new cluster parameter dashboard_settings and for the new section dashboard. Tests were also adapted concerning the template rendering.

Signed-off-by: Alexandre Gobeaux <gobeaa@amazon.com>