Use AZ IDs instead of AZ names in AZ allowed list
AZ IDs (e.g., 'use1-az6') refer to the same underlying AZ regardless of
account, while AZ names may refer to different underling AZs depending
on the account using them. This is described here:
https://docs.aws.amazon.com/ram/latest/userguide/working-with-az-ids.html.

This change enables the use of AZ IDs to specify which AZs within a
given region should be used for testing. This is so tests are always run
in the same set of AZs regardless of the account running them.

The AZ name to AZ ID transformations made in this change were made by
assuming ReadOnly credentials for the account used to run tests,
creating a region-to-AZ name-to-AZ ID mapping for this account using the
newly added `get_az_id_to_az_name_map` function, and then modifying
conftest.py based on it. Here's the mapping if it's of interest:

```json
{
  "us-east-1": {
    "use1-az6": "us-east-1a",
    "use1-az1": "us-east-1b",
    "use1-az2": "us-east-1c",
    "use1-az4": "us-east-1d",
    "use1-az3": "us-east-1e",
    "use1-az5": "us-east-1f"
  },
  "us-east-2": {
    "use2-az1": "us-east-2a",
    "use2-az2": "us-east-2b",
    "use2-az3": "us-east-2c"
  },
  "us-west-2": {
    "usw2-az2": "us-west-2a",
    "usw2-az1": "us-west-2b",
    "usw2-az3": "us-west-2c",
    "usw2-az4": "us-west-2d"
  },
  "ap-southeast-2": {
    "apse2-az3": "ap-southeast-2a",
    "apse2-az1": "ap-southeast-2b",
    "apse2-az2": "ap-southeast-2c"
  },
  "ap-northeast-1": {
    "apne1-az4": "ap-northeast-1a",
    "apne1-az1": "ap-northeast-1c",
    "apne1-az2": "ap-northeast-1d"
  },
  "ap-northeast-2": {
    "apne2-az1": "ap-northeast-2a",
    "apne2-az2": "ap-northeast-2b",
    "apne2-az3": "ap-northeast-2c",
    "apne2-az4": "ap-northeast-2d"
  },
  "ap-southeast-1": {
    "apse1-az2": "ap-southeast-1a",
    "apse1-az1": "ap-southeast-1b",
    "apse1-az3": "ap-southeast-1c"
  },
  "ap-south-1": {
    "aps1-az1": "ap-south-1a",
    "aps1-az3": "ap-south-1b",
    "aps1-az2": "ap-south-1c"
  },
  "sa-east-1": {
    "sae1-az1": "sa-east-1a",
    "sae1-az2": "sa-east-1b",
    "sae1-az3": "sa-east-1c"
  },
  "eu-west-1": {
    "euw1-az1": "eu-west-1a",
    "euw1-az2": "eu-west-1b",
    "euw1-az3": "eu-west-1c"
  }
}
```

Signed-off-by: Tim Lane <tilne@amazon.com>