Handle 'optimal' when validating CCE error message parsing
In order to make sure the supported instance types and families parsed
from the Batch's CCE are sane, they are compared to the list of all
known instance types and families returned by EC2's
DescribeInstanceTypeOfferings API.

This check was not accounting for the fact that Batch users can specify
an instance type/family of `optimal`. This change fixes that.

Signed-off-by: Tim Lane <tilne@amazon.com>