Include 'optimal' in fallback value for valid Batch compute instance types
**Summary**

This commit ensures that `optimal` is not determined to be an invalid
value when the CLI can't guarantee that the set of instance types and
families parsed from a CCE error message

**Details**

As of 6c4233e8, supported values for the `compute_instance_type` are
parsed from the error message resulting from calling Batch's
CreateComputeEnvironment API with a nonexistent instance type.

In order to ensure that the values parsed from this error message aren't
invalid (due to, for example, a modification to the CCE error message
format), they are validated against the instance types and families
returned by EC2's DescribeInstanceTypeOfferings API. In the event that
any of the instance types and families parsed from the CCE error message
aren't in the set returned by the EC2 API, the code assumes that there's
an issue with the CCE error message parsing, and returns the set
returned by the EC2 API instead.

This change adds `optimal` to the set of instance types and families
returned by the EC2 APIs that is used for both validating the values
parsed from the CCE error message and as the fallback value when this
validation fails. This is essentially an extension of the change made in
e7099774, which accounted for 'optimal' when validating the parsed
values but didn't include it in the fallback value.

Signed-off-by: Tim Lane <tilne@amazon.com>