Add integration test for tag propagation
The test verifies that the tags passed via the config file, via the
command line, and added by default are all present on head and compute
nodes, the main CFN stack, the head node substack, the compute fleet
substack, and the shared directory EBS volume.

Though the bug that motivated this test only applies to HIT clusters,
the test supports Batch and SIT clusters as well.

Signed-off-by: Tim Lane <tilne@amazon.com>