Fix extra json "cluster" vs "cfncluster" in pcluster update (#2129)
This is a compatibility fix. We have been allowing user to either use "cluster" and "cfncluster" in extra json field. And we are encouraging user to use "cluster". This commit will make sure we use "cluster" everywhere in CLI and only transform it to "cfncluster" when we want to send it to CloudFormation. Notice: if auto_refresh == False, you may get cfncluster when reading from a config file.

Signed-off-by: Hanwen <hanwenli@amazon.com>