Add Head Node logs only if CW Logs are enabled and add tests
The logs file are present in CloudWatch only if the cw_log feature is enabled.

I also removed useless conditions in logs when supported by all OSes.

## Unit Tests

Added more unit tests for dashboard substack rendering:
Now we're testing all the conditions that change the rendered template.
* CloudWatch logs: if not enabled we don't add the logs widgets
* DCV: conditional log files
* EBS and RAID volume type: conditional EC2 metrics
* EFS performance mode: conditional EC2 metric
* Dashboard enabled: global condition for the substack

## Integration Tests
Extended `test_runner` to specify a custom dashboard template url
This new parameter is required, because the dashboard template is
rendered starting from a source template that is not yet available
in the official ParallelCluster S3 buckets.

Added integration tests for the following use cases:
* dashboard created and contain Head Node Logs
* dashboard created without Head Node Logs (CW logs feature disabled)
* dashboard not created

Signed-off-by: Enrico Usai <usai@amazon.com>