Use EBS section keys/labels when initializing from CFN
Previously, parameters for EBS sections referred to by an EBS settings
config from a cluster section were being initialized with section keys
and labels for the cluster section instead of the referred EBS section.
This led to issues when these parameters tried to look the values of
other parameters rom their same section.

Signed-off-by: Tim Lane <tilne@amazon.com>