Enable minimal pylint checks and fix errors
This commit adds a pylint check to the checks run for the code-linters
tox environment. Currently the check only looks for function calls which
are missing required arguments. The motivation for enabling this
particular check is the bug that was introduced by 65cd4db1, where a
required arg was added to the function _upload_hit_resources, but the
usage in update.py was not modified accordingly.

Signed-off-by: Tim Lane <tilne@amazon.com>