Add a simple caching mechanism for functions
This commit introduces a very simple caching mechanism based on a decorator that can be used to wrap functions that we want to cache.
The goal is to reduce the number of (duplicated) boto3 calls for actions like `describe_instance_types` that can be called multiple times for the same instance type.

Signed-off-by: ddeidda <ddeidda@amazon.com>