Fail gracefully when validating invalid bucket name
Fix to avoid printing stacktrace  when an empty or invalid value is set for `cluster_resource_bucket` parameter.

NOTE: unit tests are still to be implemented for this validator; this will require a bit of mocking, better to be handled as a separate commit.

Manual tests performed:

pcluster create / update with following values:

`cluster_resource_bucket = `
`cluster_resource_bucket = abc#_1!`

Signed-off-by: ddeidda <ddeidda@amazon.com>