Fix pcluster configure internal refresh flow
Before this patch the code was loading the configuration file by initializing `PclusterConfig`
with `auto_refresh=True`. This setting means that the internal parameters must be recalculated
every time there is a change in one of the internal sections.

The issue was that the calculation of internal parameters was using the values from the
initial config file and after that, the `pcluster configure` workflow wasn't changing
internal sections (parameters only) so the auto-refresh wasn't triggered anymore.

For this reason there was a piece of code to explicitly reload the config file again after
changing the region. We don't need it anymore because I'm instead explicitly calling the
`refresh` method.

Then, in the `no_awsbatch_*_with_config_file" test we were wrongly expecting the `AWSBatchFullAccess`
policy with `sge` scheduler(!). The test was clearly wrong and was hiding a wrong behaviour.

Now we're executing the `refresh` command right after collecting all the inputs from
the users, so the `additional_iam_policies` will contain the values related to the right
scheduler (`sge`) and not the original one (`awsbatch`).

We're also setting the `auto_refresh=True` before calling the `HitConverter.convert`
because this method can change some internal sections so the internal parameters must be
updated accordingly.

This commit also fixes cases on which `refresh` method was called with default/initial values
(e.g. t2.micro as instance type in regions on which this instance doesn't exist).
Right now we're instead validating the values coming from the user.

Signed-off-by: Enrico Usai <usai@amazon.com>