Enable rollback data generation for arbitrary S3 files
Previously generating rollback data was only supported by the modified
code for S3 keys related to the config files used to allow list certain
features/instance types. This enables the same code to be used for
generating rollback data for other S3 files. The motivation for this is
to enable the code to be used to generate rollback data for dependencies
to be hosted in S3.

Signed-off-by: Tim Lane <tilne@amazon.com>