Remove explicit call to set/unset credentials in CfnStacksFactory.create_stack
set/unset credentials are called by the fixture setup_sts_credentials which has the autouse property set to true

The set credentials is moved to the _create_vpc_stack method
A check to avoid credentials being set twice is implemented
Also, add use of fixture setup_sts_credentials to networking test, so that operation done against the stack are enclosed in lifecycle of the credentials.

What the change solve? Avoid to setting credentials when they are already set

Signed-off-by: Luca Carrogu <carrogu@amazon.com>