Fix EFA OS architecture validator
The efa_os_arch_validator has been introduced in the 2.10.1 version, with patch 2290.

The configuration files with the following combination of settings are affected by the issue:
`scheduler=slurm`, `base_os=centos8`, `queue_settings` specified in the file
and `enable_efa` not set in the `cluster` section or set to `false` in the `queue` section.

The error message is coming from the validation of the `enable_efa` parameter in the `queue` section,
because this parameter has a default value so the validator is always executed.

The code of the validator is not checking if the value is set to true so it is failing in any case.

## Tests

Added two unit test to verify the cases on which there is no value for `enable_efa` in the `cluster` section.

Signed-off-by: Enrico Usai <usai@amazon.com>