Allow skipping tests when cluster creation fails
Sometimes integration tests results are "polluted" by cluster creation failures, raising false alarms on the specific features that should have been the target of the tests.
With this commit we allow tests to be skipped, instead of making them fail, when the cluster creation is not part of the real test we want to perform. The reason of the skip will be logged for deeper analysis.

Signed-off-by: ddeidda <ddeidda@amazon.com>