Collapse "config" layer into the "model" one
The `config` folder will be replaced by the `models` one
and all the `*Config` classes will became `*Resource` classes.

The idea is that by parsing the items from the configuration file,
through the schema, we'll obtain directly the core items of the model.
We don't need a separate layer for the config.

We're instead moving the logic to manage the model classes (e.g. Cluster)
in the ClusterManager class that will permit to execute all the actions
on the model classes.

## Other changes
* Moved Cluster Stack CDK code in a separate `cluster_stack` module
* Aligned tests to the previous changes
* Removed useless storage module. We'll re-add it when/if needed.

Signed-off-by: Enrico Usai <usai@amazon.com>