3.0 CLI: Apply MarkedValue to fields except of type Resource or List
1. To append more info to basic classes such as str, int , bool, float, we create a class MarkedValue to wrap the value and add `implied` and `valid` fields. The reason of not wrapping object of `Resource` to avoid writing lnng sentence to access a value, e.g `cluster.head_node.value.networking.value.subnet.value`.
2. Utilize `valid` field to build the dependencies of ebs validators. Although the approach is still under discussion, I'd include it in the PR to show what can we do with marked values.
3. Delete old ebs validators after migration.

Major Concern: This approach adds noticeable overhead when when accessing value (`.value` has to be used).

The reason of not using a method in the base Resource class to require validity of all input args of a validator is that we are still dealing with the ambiguity between separated EBS validators vs centrally orchestrated FSx validators.