Register validators right before resource validation
The previous approach of registering the validators in the initialization
of the Resource wasn't considering the possibility to change the model attributes.
The registered validators were validating only the attributes present at init time.

With the new approach the validator are registered just before calling the validation
so we'll have the validators associated to the current status of the model.
For this reason we don't need anymore the Dynamic param because we don't need to pass
a reference to the validator but we can simply pass the property with the
right value it has in that specific moment.

By removing the DynamicParam concept we're simplifying the tests
and the usage of the model.

Signed-off-by: Enrico Usai <usai@amazon.com>