Add CFN conversion via CDK of Main Stack Resources
This commit adds support for the creation of all the CFN resources contained in the legacy main CloudFormation template.

The following differences will be found between this new set of instances and the old ones:

- 1. ASG related resources are not longer required because only Slurm clusters are supported, so they have not been included in the new stack
- The configuration structure has changed (e.g. iam roles for head and compute nodes are now declared separately).
- The dynamic creation logic allows to unify resources that were managed separately in the json template (e.g. S3Read/ReadWritePolicies are now unified under the new S3AccessPolicies).

NOTE: Resources managed from substacks in the old Cfn Template have not been included in this commit.
Signed-off-by: ddeidda <ddeidda@amazon.com>