Amend/Add Schema: compute resource, custom actions, iam, image, dns
1. Add `Name` field to compute resource
2. Move custom actions to headnode and queue of slurm and headnode of awsbatch
3. Add IAM support in headnode and queue
4. Add image support in headnode and queue
5. Add DNS settings support in slurm settings

ToDo
Although IAM section has a minor difference depending on if it is on global level or lower level, I am using the same classes to make the code concise. The same idea applies to image section. We need to revisit to see whether we can change the design of the configuration file or we have to write more code for it.