Fix instance type sanity checks
This commmit replaces the latest linux image id with the actual cluster ami in dryrun run-instances calls performed during sanity checks.

This new behaviour fixes in particular the issue occurring when the architecture of the instance types is different from the one of the latest linux AMI, and provides a more reliable check when custom amis are used.

Signed-off-by: ddeidda <ddeidda@amazon.com>