Fix pcluster update behavior with additional instance types data
Additional instance types data must be loaded not only when loading the configuration from file, but also from CloudFormation in order to allow the refresh operations to work correctly.

This commit also marks as ignored the update policy of the `instance_types_data` parameter to avoid changes being displayed just because of the json data re-ordering.

Signed-off-by: ddeidda <ddeidda@amazon.com>