Integ-tests: Fix test multiple jobs submission
In pcluster2, we had 1 initial dynamic compute node. We removed the initial count in the configuration file in pcluster3. To closely imitate the original test behaviour, I change the test code to use a sleep job to start a dynamic compute node. Of course, in the future, we can consider rewriting the test

Signed-off-by: Hanwen <hanwenli@amazon.com>