Add lazy initialization and reload on region change to AWSApi
boto3 clients are not initialized on first usage rather than when the AWSApi instance is created. This speeds up
execution for those operations that do not require all clients to be created. Also added some logic to invalidate all
clients when the AWS_DEFAULT_REGION is updated. This happens when the Lambda environment is reused across multiple API
invocations and the region differs across multiple requests.

Signed-off-by: Francesco De Martino <fdm@amazon.com>