Add bucket name and artifact dir to cfn tag
Bucket name and artifact dir were stored in the outputs of Cloudformation stack. However, if the stack creation fails, the output is empty. Therefore, this commit will store them in the tags of the stack. Moreover, the infos in outputs are kept because there are tests relying on the output

Signed-off-by: Hanwen <hanwenli@amazon.com>