Allow to set up a private API
We want to give the customer the ability to configure a private API. At
the moment we are not interested in configuring the lambda to access
resources in a VPC with no internet access. In other words, the goal is
simply to enable someone to call the API from an EC2 instance running in
an isolated network.

This patch allows the customer to optionally specify a VPC endpoint id,
in which case CFN will set up a private API associating the VPC
endpoint. Then the customer will be able to call the API from an
instance in an isolated subnet in the VPC that can contact the endpoint
by calling one of the following URLs:

1. `https://{restapi-id}.execute-api.{region}.amazonaws.com/prod` (only if private DNS is enabled)
2. `https://{rest-api-id}-{vpce-id}.execute-api.{region}.amazonaws.com/prod`