Change resource names to handle stack update
Before the stack update operation could fail for reasons tied to the resource names.

The custom resource policy to delete old Docker images applies to the created private
ECR repository. However, if we change the ParallelCluster version, then the repository
is recreated and its ARN changes. This patch creates a private repository which contains
only the stack id and tags it with the ParallelCluster version instead. Changes to the tag
will not replace the stack, thus its ARN will not change.

The container recipe is replaced in place if its definition changes, however since its ARN does
not change if, for example, we update the ParallelCluster version, then the update would fail, as
replacement works by first creating the new resource and then deleting the old one, and the stack
update would fail when we try to create a new recipe with the same ARN of the old one. By encoding
the ParallelCluster version in the name we avoid this problem, as the new recipe would have a
different ARN.

Changes in the container recipe will not trigger a replacement of the Image resource (this is different
from the behavior of the Image resource in case of changes to an image recipe). In order to trigger a
replacement this patch also encodes the ParallelCluster version in the name of the infrastructure configuration,
as changes to this latter resource do trigger a replacement of the Image resource.