Remove parallelcluster- stack prefix
Add stack id to the path of IAM roles for slurm cluster. Because path has a higher limit (256 characters) than role name (64 characters), it is safe to put the whole stack arn in the path. We still need to discuss whether we want to add parallelcluster-prefix to the role name which can facilitate the scoping down of user policies.

This commit also modify the changelog unrelated to the code change of this commit

Signed-off-by: Hanwen <hanwenli@amazon.com>