Add integration test for Infrastructure API CFN template
This patch adds an infrastructure tests validating that when creating a
stack with the Infrastructure API CloudFormation template using the
default parameters the APIGateway API backed by Lambda is correctly set
up. The test requires inputting two parameters, the URI of the published
Docker image for the ParallelCluster Lambda and the S3 URI of the API
definition.