Implement ListImages API
This patch implements the ListImages API. It takes as a required
argument an imageStatus, which can be AVAILABLE, PENDING or FAILED, and
returns a summary compiled with the information about the available
instances (EC2 describe call) or the available stacks (we filter the
Boto3 call result as the CloudFormation describe_stacks operation unfortunately
does not support filters). The describe_stacks call is the only paginated
one, for the AVAILABLE instances we do a filtered EC2 describe_instances
call that gets the "available" images, and return None as a nextToken.