Implement DeleteImage API
This patch implements the DeleteImage API. We create an ImageBuilder
class with the given image id and retrieve the associated image or
stack, then we delete the image and, in case we only had the stack, we
retrieve it with another ImageBuilder object (the stack is initialized
only once, and if we tried to retrieve it after calling the delete
operation we might be unable to describe it as it might have already
transitioned to the DELETE_COMPLETE state).

One thing to note here is that the ImageBuilder.delete() call might fail
raising an ImageBuilderActionError, which is surfaced to the caller of
the DeleteImage API as a 500 with the message: "Unexpected fatal
exception. Please look at the application logs for details on the
encountered failure."