Implement DescribeOfficialImages
This patch implements the DescribeOfficialImages API. We search for
images owned by "amazon" with the correct name prefix
(aws-parallelcluster-...), optionally taking version, os and
architecture parameters which are used to narrow down the search by
constructing an appropriate prefix pattern. Note that in the AmiInfo
structures in the response we need to parse the image name in order to
retrieve the correct information about the operating system, as tags are
not visible to accounts different from the one that owns the image
(these are official images, not images that have been built by the
customer), and the PlatformDetails section of the EC2 describe_images
response is not specific enough (i.e. it generically identifies the
images as a Linux image, for example, but won't distinguish between
CentOS, Ubuntu, etc.).

Note also that we use a map to convert the specified os into an
appropriate suffix. In case the os is not present, we do not throw an
exception, but rather build a name prefix with an empty string and
return an empty list (this is to be coherent with the behavior when a
non existing version or architecture is specified).