In DeleteImage response set state to DELETE_IN_PROGRESS
Before, we set the ImageBuilderStatus to DELETE_COMPLETE when deleting
an available image, and tried to get the CloudFormation stack status
when deleting a stack, setting it to DELETE_COMPLETE if we did not find
and setting the ImageBuilderStatus to the corresponding state (the
reasoning here is that the deletion of an available image is synchronous
and so we would have had exceptions in case of failure, while the
deletion of a stack is asynchronous).

We have decided to statically set the ImageBuilderStatus and
CloudFormation stack status to DELETE_IN_PROGRESS in the response.