Implement DescribeImage
This patch implements the DescribeImage API. We initialize an
imageBuilder object, and try to retrieve the available image or, in case
it is not yet available, the CloudFormation stack (eventually with its
associated ImageBuilder image information). In case there are no
available images or stacks associated with the provided image id, we
throw a 404 to indicate that we did not find a matching image.