Return more specific errors in case of create image request validation failure
This patch returns more detailed failures in case of issues during a
create image request. Specifically, it appropriately converts conflicts
(i.e. when creating an image with the same id as an existing
image/stack) exceptions, as well as bad request exceptions due to
validation failures (in this case, the failure level, message and
validator are returned in the thrown exception).