DescribeOfficialImages: use current PC version
This patch removes the version filter and adds it to the response. The
version used to build the name prefix is the current PC version, we
don't allow the caller to specify an arbitrary version. In the
controller we validate the optionally specified os and architecture
against the allowed values specified in the constants.py file.