Implement UpdateCluster API
This patch implements the UpdateCluster API by calling the methods
exposed by the Cluster business object. Things to note:

1. We throw a BadRequest execption in case of an empty change set,
unless the caller specified force=True. Note that this change set does
not take into account changes to parameters with print_succeded = False.
As far as we can tell at the moment this can happen only for SUCCEDED
check result with policies IGNORED or READ_ONLY_RESOURCE_BUCKET. In the
future we plan to make these be UNSUPPORTED.
2. The validation of the patch happens after the validation of the
configuration. We do not return suppressed validation configuration
errors in the BadRequest caused by a failed during patch validation, as
we do not want to confuse the customer (i.e. the BadRequest was not
caused by the configuration validation erros they suppressed). We do
return the suppressed configuration validation errors in the response in
the case of a successful request.