Simplify list comparison logic
When computing the config changes, we can simplify the comparison logic
for lists. Namely, we always have an update key, and unvisited base
config subsections are surely not present in the target config, as
otherwise we would have already visited them when evaluating the target
sections present also in the base section in the previous step. For this
reason, we can remove those parts of the code, since they are never
executed.