Remove schema manipulation for SharedStorageSchema sections to fix regression on update checks
We needed a way to differentiate different storage schemas
and we were using an internal manipulation with pre_load to create internal fields
to be used to validate the different schemas.

It was causing problem during the check of the changes in the update
because we were cycling for all the schema variables and these internally created
fields were not present in the schema.

This bug was not discovered by config_patch tests because we were using the old
schema with `SharedStorage > Ebs` structure rather than the current schema on
which "Ebs" is not a subfolder but just an attribute of `SharedStorage > StorageType:Ebs`

In this case we don't really need pre_load, post_dump actions to create internal fields
to differentiate the storage schemas. We can simply accept both Settings fields
with different schemas and just differentiate the initialization according to the storage type.

Added unit tests for the new validators.

Signed-off-by: Enrico Usai <usai@amazon.com>