Remove schema manipulation for Scheduling section to fix regression on update checks
We needed a way to differentiate Slurm queues schema from Awsbatch queues schema
and we were using an internal manipulation with pre_load to create internal fields
to be used to validate the different schemas.

It was causing problem during the check of the changes during an update
because we were cycling for all the schema variables and these internally created
fields were not present in the schema.
This bug was not discovered by config_patch tests because we were using the old
schema with `Scheduling > Slurm` structure rather than the current schema on
which "slurm" is not a subfolder but just an attribute of `Scheduling > Scheduler:Slurm`

Now instead of doing pre_load, post_dump actions to create internal fields
to differentiate the scheduling schemas are accepting both SlurmQueues and AwsBatchQueues
with different schemas and just differentiate the initialization according to the scheduler.

This is aligned with Slurm/AwsBatchSettings or Efs/Ebs/FsxLustreSettings,
this simplify validation because we don't need to post-process failures from marshmallow
to restore the right section name used by the customer.

## Tests
* Fixed unit tests for config_patch by using schema load and being sure we're using an accepted schema
* Added unit test for a change in the EbsSettings fields
* Updated configuration files in unit tests expected outputs and integration tests
* Added multiple unit tests for SchedulingSchema and SharedStorageSchema
* Updated integration tests configuration files

Signed-off-by: Enrico Usai <usai@amazon.com>