Do not convert generic Exceptions at the API controller level.
We want to avoid to show generic exceptions' messages in the error
returned to the customer. This patch makes the decorator in the
controllers simply raise the generic exceptions and we then rely on the
conversion in the flask app that produces a error with a generic
message, telling the customer to look at the logs for more details.