Default to DEFAULT_MAX_COUNT in UpdatePolicy.MAX_COUNT condition checker
In the UpdatePolicy.MAX_COUNT condition checker we want to check that
either the fleet is stopped, or the new max count is bigger or equal to
the old one (because shrinking queues requires stopping the compute
fleet). If the old configuration and/or the new one do not specify a max
count, we default to DEFAULT_MAX_COUNT in the condition checker.