Add parameter DeleteOnTermination in head node root volume and DeletionPolicy in Shared EBS
DeletionPolicy for shared EFS and FSx is not added because: keeping EFS or FSx will lead to stack deletion failure because a network interface is attached to them and the security group we created is associated with the network interface. The easiest solution would be keep the compute security group too. However, when the user try to reuse the file system, they need to attach another proper security group to be able to access.

Signed-off-by: Hanwen <hanwenli@amazon.com>