Add integration test for cluster operation APIs
For the API test we want to cover just a minimal subset of feature
while we will rely on CLI tests for full coverage (being the CLI a
wrapper on top of the API). For the cluster APIs we have 2 test cases,
one for Slurm and one for AWS Batch.

The tests' workflow starts by creating a cluster, waiting until it has
been created and then performing a list and describe, verifying that it
can be found. Then we do a dry-run update, verifying that an eventual
update would have succeeded, and lastly we delete the cluster, verifying
that it does not appear anymore in the listing. For the moment we do not
have cluster instance or compute fleet API calls, when these APIs are
ready we will add them in the tests.