Use api_server_factory to create stack for api infrastructure integration tests
Before, these tests referenced a local CloudFormation template to create
the stack to set up the ParallelCluster API infrastructure. This patch
uses the `apy_server_factory` fixture that we also use to instantiate
the api_client in the other API tests to create the infrastructure,
removing the dependency from the local template file.