Extend cluster APIs integration tests
This patch completes the cluster APIs integration tests. In particular,
it adds support for Cluster Instance and Compute Fleet APIs, and makes
the check on the DeleteCluster check more efficient by making a
DescribeClusters call instead of a ListClusters.

One thing to note is that in the UpdateComputeFleet test we have some
differences between the Slurm and AWS Batch case, in particular the state
names are different, and in the latter case the nodes are not completely
removed, instead a number of instances equal to MinvCpus is kept, so we
cannot wait until all the existing compute nodes have been terminated,
and we simply check the status in that particular case.